+++ R/regplot.r
@@ -1,0 +1,2 @@
+regplot <- function(x, ...)
+   UseMethod("regplot")